- Version 7 of Guzzle has a bunch of security advisories: https://packagist.org/packages/guzzlehttp/guzzle/advisories?version=4099679 - The Monolog and Guzzle dependencies are quite strict, both are widely used and the strictness will likely create a "dependency hell" for many users of your SDK. As far as i could see the functionality you use is identical in both versions so i like to ask you to support both. The alternative is probably that users start forking and lose their ability to receive your updates.
Not really sure how your test suite is supposed to work but it seems like the path as specified in "autoload-dev" is incorrect. I think it should be "src/Tests/". Not sure if this was causing the issues that i ran into but when i attempted to run the code it was unable to find my .env file. Instead of looking for what caused it i just copied the .env file over to a bunch of directories until it did work. Maybe this was what caused it.